doc: Merge Mousewheel interaction into Base interaction
