+++ apps/insight-viewer-demo/src/containers/Interaction/Base.tsx
@@ -1,19 +1,40 @@
-import { Box } from '@chakra-ui/react'
+import { Box, Text } from '@chakra-ui/react'
 import Viewer, {
   useInteraction,
   useViewport,
+  useMultiframe,
   Interaction,
+  Wheel,
 } from '@lunit/insight-viewer'
 import CodeBlock from '../../components/CodeBlock'
 import Control from './Control'
+import WheelControl from './Control/Wheel'
 import OverlayLayer from '../../components/OverlayLayer'
 import CustomProgress from '../../components/CustomProgress'
 import { BASE_CODE } from './Code'
+import Canvas from './Canvas'
 
-const IMAGE_ID =
-  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000011.dcm'
+const IMAGES = [
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000000.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000001.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000002.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000003.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000004.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000005.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000006.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000007.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000008.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000009.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000010.dcm',
+]
+
+const MIN_FRAME = 0
+const MAX_FRAME = IMAGES.length - 1
+const MIN_SCALE = 0.178
+const MAX_SCALE = 3
 
 export default function App(): JSX.Element {
+  const { image, frame, setFrame } = useMultiframe(IMAGES)
   const { interaction, setInteraction } = useInteraction()
   const { viewport, setViewport } = useViewport()
 
@@ -26,18 +47,54 @@ export default function App(): JSX.Element {
     }
   }
 
+  const handleFrame: Wheel = (_, deltaY) => {
+    if (deltaY !== 0)
+      setFrame(prev =>
+        Math.min(Math.max(prev + (deltaY > 0 ? 1 : -1), MIN_FRAME), MAX_FRAME)
+      )
+  }
+
+  const handleZoom: Wheel = (_, deltaY) => {
+    if (deltaY !== 0)
+      setViewport(prev => ({
+        ...prev,
+        scale: Math.min(
+          Math.max(prev.scale + (deltaY > 0 ? 0.25 : -0.25), MIN_SCALE),
+          MAX_SCALE
+        ),
+      }))
+  }
+
+  const handler = {
+    frame: handleFrame,
+    zoom: handleZoom,
+  }
+
+  function handleWheel(value: string) {
+    setInteraction((prev: Interaction) => ({
+      ...prev,
+      mouseWheel:
+        value === 'none' ? undefined : handler[value as keyof typeof handler],
+    }))
+  }
+
   return (
     <Box w={700}>
       <Control onChange={handleChange} />
+      <WheelControl onChange={handleWheel} />
+      <Box mb={6}>
+        <Text>frame: {frame}</Text>
+      </Box>
       <Box w={500} h={500}>
         <Viewer.Dicom
-          imageId={IMAGE_ID}
+          imageId={image}
           interaction={interaction}
-          viewport={viewport}
           onViewportChange={setViewport}
+          viewport={viewport}
           Progress={CustomProgress}
         >
           <OverlayLayer viewport={viewport} />
+          <Canvas viewport={viewport} />
         </Viewer.Dicom>
       </Box>
       <Box w={900}>
